Lab5 in Symfony

+++ Symfony/my_project_directory/src/Controller/AuthorController.php
@@ -9,11 +9,13 @@ use App\Repository\AuthorRepository;
 use App\Service\PaginationService;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\Finder\Exception\AccessDeniedException;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 
-#[Route('/author')]
+#[Route('/api/author')]
 final class AuthorController extends AbstractController
 {
     #[Route(name: 'app_author_index', methods: ['GET', 'POST'])]
@@ -21,10 +23,13 @@ final class AuthorController extends AbstractController
                           AuthorRepository $authorRepository,
                           PaginationService $paginationService): Response
     {
+        if (!$this->isGranted('ROLE_USER')) {
+            throw $this->createAccessDeniedException();
+        }
+
         $page = $request->get('page', 1);
         $itemsPerPage = $request->get('itemsPerPage', 2);
 
-        // Форма фільтрації
         $form = $this->createForm(AuthorFilterType::class);
         $form->handleRequest($request);
         $filters = $form->getData() ?? [];
@@ -76,6 +81,11 @@ final class AuthorController extends AbstractController
     #[Route('/new', name: 'app_author_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN') &&
+            !$this->isGranted('ROLE_MANAGER')) {
+            throw $this->createAccessDeniedException();
+        }
+
         $author = new Author();
         $form = $this->createForm(AuthorForm::class, $author);
         $form->handleRequest($request);
@@ -96,14 +106,22 @@ final class AuthorController extends AbstractController
     #[Route('/{id}', name: 'app_author_show', methods: ['GET'])]
     public function show(Author $author): Response
     {
+        if (!$this->isGranted('ROLE_USER')) {
+            throw $this->createAccessDeniedException();
+        }
         return $this->render('author/show.html.twig', [
             'author' => $author,
         ]);
     }
 
     #[Route('/{id}/edit', name: 'app_author_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, Author $author, EntityManagerInterface $entityManager): Response
+    public function edit(Request $request, Author $author, EntityManagerInterface $entityManager, AuthorizationCheckerInterface $authChecker): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN') &&
+            !$this->isGranted('ROLE_MANAGER')) {
+            throw $this->createAccessDeniedException();
+        }
+
         $form = $this->createForm(AuthorForm::class, $author);
         $form->handleRequest($request);
 
@@ -122,6 +140,11 @@ final class AuthorController extends AbstractController
     #[Route('/{id}', name: 'app_author_delete', methods: ['POST'])]
     public function delete(Request $request, Author $author, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN'))
+        {
+            throw $this->createAccessDeniedException();
+        }
+
         if ($this->isCsrfTokenValid('delete'.$author->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($author);
             $entityManager->flush();
